move functions to branch.py
