+++ BNS_JT/tests/test_branch.py
@@ -2,6 +2,7 @@ import pandas as pd
 import numpy as np
 import networkx as nx
 import pdb
+import copy
 import time
 import pytest
 import json
@@ -14,6 +15,13 @@ HOME = Path(__file__).absolute().parent
 PROJ = HOME.joinpath('../../')
 np.set_printoptions(precision=3)
 #pd.set_option.display_precision = 3
+
+def compare_list_of_sets(a, b):
+
+    return set([tuple(x) for x in a]) == set([tuple(x) for x in b])
+
+
+
 
 def test_run_bnb():
     info = {'path': [[2], [3, 1]],
@@ -685,4 +693,351 @@ def test_get_set_branches_sf():
 
     arcs = branch.get_arcs_given_bstar(bstar, path_time_idx, arc_cond)
     sb = branch.get_set_of_branches(bstar, arcs, path_time_idx, arc_cond)
+
+
+def test_approx_branch_prob():
+
+    d = {f'e{i}': 0 for i in range(1, 7)}
+    u = {f'e{i}': 2 for i in range(1, 7)}
+    br = branch.Branch(d, u, 's', 's')
+
+    p1 = {0: 0.05, 1: 0.15, 2: 0.80}
+    p2 = {0: 0.01, 1: 0.09, 2: 0.90}
+    p = {'e1': p1, 'e2': p1, 'e3': p1,
+         'e4': p2, 'e5': p2, 'e6': p2}
+
+    br.approx_prob(p)
+    assert br.p == 1.0
+
+    d = {'e1': 2, 'e2': 0, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    u = {f'e{i}': 2 for i in range(1, 7)}
+    br = branch.Branch(d, u, 's', 's')
+
+    br.approx_prob(p)
+    assert pytest.approx(br.p) == 0.80**2*0.90**3
+
+    d = {'e1': 0, 'e2': 1, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    u = {f'e{i}': 2 for i in range(1, 7)}
+    br = branch.Branch(d, u, 's', 's')
+
+    br.approx_prob(p)
+    assert pytest.approx(br.p) == 0.95
+
+    d = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    u = {'e1': 2, 'e2': 0, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    br = branch.Branch(d, u, 's', 's')
+
+    br.approx_prob(p)
+    assert pytest.approx(br.p) == 0.05
+
+
+def test_approx_joint_prob_compat_rules():
+
+    p1 = {0: 0.05, 1: 0.15, 2: 0.80}
+    p2 = {0: 0.01, 1: 0.09, 2: 0.90}
+
+    p = {'e1': p1, 'e2': p1, 'e3': p1,
+         'e4': p2, 'e5': p2, 'e6': p2}
+
+    d = {f'e{i}': 0 for i in range(1, 7)}
+    u = {f'e{i}': 2 for i in range(1, 7)}
+    br = branch.Branch(d, u, 's', 's', 1.0)
+    rule = {'e2': 2, 'e5': 2}
+    rule_st = 's'
+
+    result = br.approx_joint_prob_compat_rule(rule, rule_st, p)
+    assert pytest.approx(result) == 0.8*0.9
+
+    rule = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    rule_st = 'f'
+
+    result = br.approx_joint_prob_compat_rule(rule, rule_st, p)
+    assert pytest.approx(result) == 0.05**3*0.01**3
+
+
+def test_get_compat_rules():
+
+    upper = {f'e{i}': 2 for i in range(1, 7)}
+    lower = {f'e{i}': 0 for i in range(1, 7)}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    rules = {'s': [], 'f': []}
+    result = br.get_compat_rules(rules)
+    assert result == rules
+
+    rules = {'s': [{'e2': 2, 'e5': 2}], 'f': []}
+    result = br.get_compat_rules(rules)
+    assert result == {'s': [{'e2': 2, 'e5': 2}], 'f': []}
+
+    rules = {'s': [{'e2': 1, 'e5': 2}], 'f': [{f'e{i}': 0 for i in range(1, 7)}]}
+    result = br.get_compat_rules(rules)
+    assert result['s'] == rules['s']
+    assert result['f'] == rules['f']
+
+    upper = {'e1': 2, 'e2': 0, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {f'e{i}': 0 for i in range(1, 7)}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    rules = {'s': [{'e2': 1, 'e5': 2}],
+             'f': [{f'e{i}': 0 for i in range(1, 7)}]}
+    result = br.get_compat_rules(rules)
+    assert result['s'] == []
+    assert result['f'] == [{'e1': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}]
+
+    upper = {'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {'e1': 2, 'e2': 0, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    rules = {'s': [{'e2': 1, 'e5': 2}, {'e1': 2, 'e3': 2, 'e5': 2},
+                   {'e2': 2, 'e4': 2, 'e6': 2}],
+             'f': [{f'e{i}': 0 for i in range(1, 7)}]}
+    result = br.get_compat_rules(rules)
+    assert result['s'] == [{'e2': 1}, {'e2': 2}]
+    assert result['f'] == []
+
+
+def test_get_c_from_br(main_sys):
+
+    G, _, _, d_varis = main_sys
+
+    varis = copy.deepcopy(d_varis)
+
+    st_br_to_cs = {'f': 0, 's': 1, 'u': 2}
+
+    # test1
+    br = branch.Branch({'e1': 1, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0},
+                       {'e1': 2, 'e2': 0, 'e3': 1, 'e4': 2, 'e5': 2, 'e6': 2}, 'f', 'f')
+
+    varis, cst = br.get_c(varis, st_br_to_cs)
+
+    assert cst.tolist() == [0, 5, 0, 3, 6, 6, 6]
+    """
+    assert compare_list_of_sets(varis['e1'].B, [{0}, {1}, {2}, {0, 1}, {0, 2}, {1, 2}, {0, 1, 2}])
+    assert compare_list_of_sets(varis['e2'].B, [{0}, {1}, {2}])
+    assert compare_list_of_sets(varis['e3'].B, [{0}, {1}, {2}, {0, 1}])
+    assert compare_list_of_sets(varis['e4'].B, [{0}, {1}, {2}, {0, 1, 2}])
+    assert compare_list_of_sets(varis['e5'].B, [{0}, {1}, {2}, {0, 1, 2}])
+    assert compare_list_of_sets(varis['e6'].B, [{0}, {1}, {2}, {0, 1, 2}])
+    """
+
+    # test2
+    # using the previous output as an input
+    br = branch.Branch({'e1': 0, 'e2': 0, 'e3': 2, 'e4': 0, 'e5': 0, 'e6': 0},
+                       {'e1': 2, 'e2': 0, 'e3': 2, 'e4': 2, 'e5': 1, 'e6': 2}, 'f', 'f')
+
+    varis, cst = br.get_c(varis, st_br_to_cs)
+    assert cst.tolist() == [0, 6, 0, 2, 6, 3, 6]
+    """
+    assert compare_list_of_sets(varis['e1'].B, [{0}, {1}, {2}, {1, 2}, {0, 1, 2}])
+    assert compare_list_of_sets(varis['e2'].B, [{0}, {1}, {2}])
+    assert compare_list_of_sets(varis['e3'].B, [{0}, {1}, {2}, {0, 1}])
+    assert compare_list_of_sets(varis['e4'].B, [{0}, {1}, {2}, {0, 1, 2}])
+    assert compare_list_of_sets(varis['e5'].B, [{0}, {1}, {2}, {0, 1, 2}, {0, 1}])
+    assert compare_list_of_sets(varis['e6'].B, [{0}, {1}, {2}, {0, 1, 2}])
+    """
+
+
+def test_get_new_branch2():
+
+    br = branch.Branch(down={'e1': 1, 'e2': 0, 'e3': 0},
+                       up={'e1': 1, 'e2': 1, 'e3': 1},
+                       down_state='u', up_state='s', p=0.9)
+    rules = {'s': [{'e1': 1, 'e2': 1}], 'f': []}
+
+    xd, xd_st = 'e2', 1
+
+    probs = {'e1': {0: 0.1, 1: 0.9},
+             'e2': {0: 0.2, 1: 0.8},
+             'e3': {0: 0.3, 1: 0.7}}
+
+    out = br.get_new_branch(rules, probs, xd, xd_st)
+    assert out.down == {'e1': 1, 'e2': 0, 'e3': 0}
+    assert out.up == {'e1': 1, 'e2': 0, 'e3': 1}
+    assert out.down_state == 'u'
+    assert out.up_state == 'u'
+    assert out.p == pytest.approx(0.9*0.2)
+
+    out = br.get_new_branch(rules, probs, xd, xd_st, up_flag=False)
+    assert out.down == {'e1': 1, 'e2': 1, 'e3': 0}
+    assert out.up == {'e1': 1, 'e2': 1, 'e3': 1}
+    assert out.down_state == 's'
+    assert out.up_state == 's'
+    assert out.p == pytest.approx(0.9*0.8)
+
+
+def test_get_new_branch1():
+
+    br = branch.Branch(down={'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0, 'e7': 0, 'e8': 0, 'e9': 0, 'e10': 0, 'e11': 0, 'e12': 0, 'e13': 0, 'e14': 0, 'e15': 0, 'e16': 0, 'e17': 0, 'e18': 0, 'e19': 0, 'e20': 0, 'e21': 0}, up={'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2, 'e7': 2, 'e8': 2, 'e9': 2, 'e10': 2, 'e11': 2, 'e12': 2, 'e13': 2, 'e14': 2, 'e15': 2, 'e16': 2, 'e17': 2, 'e18': 2, 'e19': 2, 'e20': 2, 'e21': 2}, down_state='u', up_state='s', p=1.0)
+
+    rules = {'s': [{'e3': 1, 'e9': 1, 'e14': 1, 'e17': 1}], 'f': []}
+    probs ={'e1': {0: 0.1163, 1: 0.0616, 2: 0.8221}, 'e2': {0: 0.1624, 1: 0.1224, 2: 0.7152}, 'e3': {0: 0.2014, 1: 0.09, 2: 0.7086}, 'e4': {0: 0.0689, 1: 0.1155, 2: 0.8156}, 'e5': {0: 0.1863, 1: 0.1366, 2: 0.6771}, 'e6': {0: 0.2244, 1: 0.0214, 2: 0.7542}, 'e7': {0: 0.222, 1: 0.1334, 2: 0.6446}, 'e8': {0: 0.1265, 1: 0.0762, 2: 0.7973}, 'e9': {0: 0.2993, 1: 0.0343, 2: 0.6664}, 'e10': {0: 0.3016, 1: 0.0813, 2: 0.6171}, 'e11': {0: 0.2385, 1: 0.0785, 2: 0.683}, 'e12': {0: 0.346, 1: 0.0269, 2: 0.6271}, 'e13': {0: 0.3512, 1: 0.0441, 2: 0.6047}, 'e14': {0: 0.0326, 1: 0.0182, 2: 0.9492}, 'e15': {0: 0.0231, 1: 0.1268, 2: 0.8501}, 'e16': {0: 0.0373, 1: 0.083, 2: 0.8797}, 'e17': {0: 0.0222, 1: 0.0192, 2: 0.9586}, 'e18': {0: 0.0052, 1: 0.0411, 2: 0.9537}, 'e19': {0: 0.3935, 1: 0.0625, 2: 0.544}, 'e20': {0: 0.0651, 1: 0.0457, 2: 0.8892}, 'e21': {0: 0.126, 1: 0.0495, 2: 0.8245}}
+
+    xd, xd_st = 'e3', 1
+
+    # up
+    br_new = br.get_new_branch(rules, probs, xd, xd_st, up_flag=True)
+    cr_new = br_new.get_compat_rules(rules)
+    assert br_new.down_state == 'u'
+    assert br_new.up_state == 'u'
+    assert br_new.p == pytest.approx(0.2014)
+    assert br_new.down == {f'e{i}': 0 for i in range(1, 22)}
+    assert br_new.up == {f'e{i}': 0 if i==3 else 2 for i in range(1, 22)}
+    assert cr_new['s'] == []
+    assert cr_new['f'] == []
+
+    # down
+    br_new = br.get_new_branch(rules, probs, xd, xd_st, up_flag=False)
+    cr_new = br_new.get_compat_rules(rules)
+    assert cr_new['s'] == [{'e9': 1, 'e14': 1, 'e17': 1}]
+    assert cr_new['f'] == []
+    assert br_new.down_state == 'u'
+    assert br_new.up_state == 's'
+    assert br_new.p == pytest.approx(0.7986)
+    assert br_new.up == {f'e{i}': 2 for i in range(1, 22)}
+    assert br_new.down == {f'e{i}': 1 if i==3 else 0 for i in range(1, 22)}
+
+
+def test_get_decomp_comp_using_probs0():
+
+    rules = {'s': [{'e2': 2, 'e5': 2}],
+            'f': [{'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}]}
+    upper = {'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    p1 = {0: 0.05, 1: 0.15, 2: 0.80}
+    p2 = {0: 0.01, 1: 0.09, 2: 0.90}
+
+    probs = {'e1': p1, 'e2': p1, 'e3': p1,
+             'e4': p2, 'e5': p2, 'e6': p2}
+
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e2', 2)
+
+    rules = {'s': [{'e2': 2, 'e5': 2}, {'e1': 2, 'e3': 2, 'e5': 2}],
+            'f': [{'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}],
+            }
+    upper = {'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e5', 2)
+
+    rules = {'s': [{'e2': 2, 'e5': 2}, {'e1': 2, 'e3': 2, 'e5': 2}, {'e2': 2, 'e4': 2, 'e6': 2}],
+            'f': [{'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}],
+            }
+    upper = {'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e2', 2)
+
+    rules = {'s': [{'e1': 2, 'e3': 2, 'e5': 2}],
+            'f': [{'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}],
+            }
+    upper = {'e1': 2, 'e2': 2, 'e3': 2, 'e4': 2, 'e5': 2, 'e6': 2}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0, 'e4': 0, 'e5': 0, 'e6': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e1', 2)
+
+
+def test_get_decomp_comp_using_probs1():
+
+    rules = {'s': [{'e1': 1, 'e2': 1}],
+            'f': []}
+    upper = {'e1': 1, 'e2': 1, 'e3': 1}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+
+    probs = {'e1': {0: 0.1,
+                    1: 0.9},
+             'e2': {0: 0.2,
+                    1: 0.8},
+             'e3': {0: 0.3,
+                    1: 0.7}}
+
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e1', 1)
+
+
+def test_get_decomp_comp_using_probs2():
+
+    rules = {'s': [{'e2': 1}],
+            'f': []}
+    upper = {'e1': 1, 'e2': 1, 'e3': 1}
+    lower = {'e1': 1, 'e2': 0, 'e3': 0}
+
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    probs = {'e1': {0: 0.1,
+                    1: 0.9},
+             'e2': {0: 0.2,
+                    1: 0.8},
+             'e3': {0: 0.3,
+                    1: 0.7}}
+
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e2', 1)
+
+
+def test_get_decomp_comp_using_probs3():
+
+    rules = {'s': [{'e1': 1, 'e2': 1}, {'e1':1, 'e3': 1}],
+             'f': []}
+    upper = {'e1': 1, 'e2': 1, 'e3': 1}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+    probs = {'e1': {0: 0.1,
+                    1: 0.9},
+             'e2': {0: 0.2,
+                    1: 0.8},
+             'e3': {0: 0.3,
+                    1: 0.7}}
+
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e1', 1)
+
+
+def test_get_decomp_comp_using_probs4():
+
+    rules = {'s': [{'e1': 1, 'e2': 1}, {'e1':1, 'e3': 1}],
+             'f': [{'e1': 0}]}
+    upper = {'e1': 1, 'e2': 1, 'e3': 1}
+    lower = {'e1': 0, 'e2': 0, 'e3': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+
+    probs = {'e1': {0: 0.1,
+                    1: 0.9},
+             'e2': {0: 0.2,
+                    1: 0.8},
+             'e3': {0: 0.3,
+                    1: 0.7}}
+
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e1', 1)
+
+
+def test_get_decomp_comp_using_probs5():
+
+    rules = {'s': [{'e2': 1}, {'e3': 1}],
+             'f': [{'e1': 0}]}
+    upper = {'e1': 1, 'e2': 1, 'e3': 1}
+    lower = {'e1': 1, 'e2': 0, 'e3': 0}
+    br = branch.Branch(lower, upper, 's', 's', 1.0)
+
+    probs = {'e1': {0: 0.1,
+                    1: 0.9},
+             'e2': {0: 0.2,
+                    1: 0.8},
+             'e3': {0: 0.3,
+                    1: 0.7}}
+    result = br.get_decomp_comp_using_probs(rules, probs)
+
+    assert result == ('e2', 1)
+
+
